refactor(spell): extract formatting extensions into DndFormatExt
