+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/spell/presentation/component/SpellCard.kt
@@ -31,15 +31,6 @@ import com.cyrillrx.rpg.spell.domain.Spell
 import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
 import rpg_companion.composeapp.generated.resources.Res
-import rpg_companion.composeapp.generated.resources.formatted_spell_school_level
-import rpg_companion.composeapp.generated.resources.school_abjuration
-import rpg_companion.composeapp.generated.resources.school_conjuration
-import rpg_companion.composeapp.generated.resources.school_divination
-import rpg_companion.composeapp.generated.resources.school_enchantment
-import rpg_companion.composeapp.generated.resources.school_evocation
-import rpg_companion.composeapp.generated.resources.school_illusion
-import rpg_companion.composeapp.generated.resources.school_necromancy
-import rpg_companion.composeapp.generated.resources.school_transmutation
 import rpg_companion.composeapp.generated.resources.spell_casting_time
 import rpg_companion.composeapp.generated.resources.spell_components
 import rpg_companion.composeapp.generated.resources.spell_duration
@@ -165,32 +156,6 @@ private fun SpellGridItem(title: String, subtitle: String, spellColor: Color) {
                 .background(spellColor),
         )
     }
-}
-
-fun Spell.getColor(): Color = Color(173, 29, 29)
-
-@Composable
-fun Spell.getFormattedSchool() =
-    stringResource(Res.string.formatted_spell_school_level, getSchool(), level)
-
-@Composable
-private fun Spell.getSchool(): String {
-    return schools.map { it.toFormattedString() }.joinToString(", ")
-}
-
-@Composable
-fun Spell.School.toFormattedString(): String {
-    val stringRes = when (this) {
-        Spell.School.ABJURATION -> Res.string.school_abjuration
-        Spell.School.CONJURATION -> Res.string.school_conjuration
-        Spell.School.DIVINATION -> Res.string.school_divination
-        Spell.School.ENCHANTMENT -> Res.string.school_enchantment
-        Spell.School.EVOCATION -> Res.string.school_evocation
-        Spell.School.ILLUSION -> Res.string.school_illusion
-        Spell.School.NECROMANCY -> Res.string.school_necromancy
-        Spell.School.TRANSMUTATION -> Res.string.school_transmutation
-    }
-    return stringResource(stringRes)
 }
 
 @Preview
